requirements.txt
# Core Django and environment handling
Django>=5.2,<6.0
python-decouple>=3.8
psycopg2-binary>=2.9

# Django REST Framework
djangorestframework>=3.15